+++ db/migrate/20160704095606_create_wallets.rb
@@ -11,8 +11,8 @@ class CreateWallets < ActiveRecord::Migration[4.2]
 
     # create all wallets
     execute <<-SQL
-      INSERT INTO wallets (user, amount, created_at, updated_at)
-      SELECT users.id, 0, #{DateTime.current.iso8601}, #{DateTime.current.iso8601}
+      INSERT INTO wallets (user_id, amount, created_at, updated_at)
+      SELECT users.id, 0, '#{DateTime.current.iso8601}', '#{DateTime.current.iso8601}'
       FROM users
     SQL
   end